Cambios temporales

+++ models.py
@@ -78,7 +78,7 @@ class Renta(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     edificio: str
     habitaciones: int
-    banos: float
+    banos: int
     lat: float
     lon: float
     mascotas: bool
